added hashtag
added # character, continues on issue #2, it maybe should be discussed whether this conventions is used in all of our projects. One potential issue is that we will have problems if we start working with data that actually contains hashtags, but on the other hand it is unlikely this kind of data would be written into ELAN files anyway?