added some punctuation characters
These alternative quotation marks do occur regularly in Komi texts. So it is a bit open question whether we want to store somehow distinctly punctuation characters that occur in our spoken data only, or if we want to have open-ended list for all punctuation characters that occur in Komi texts more generally. Probably it has to be somehow distinguished which are accepted where... This relates to issue #2.